first pass at breaking utilities out into separate files for easier test creation
