+++ src/components/search-modules/LeadOrganization/LeadOrganization.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import Fieldset from '../../atomic/Fieldset';
 import { Autocomplete } from '../../atomic';
 import { searchLeadOrg } from '../../../store/actions';
-import { matchItemToTerm, sortItems } from '../../../utilities/utilities';
+import { matchItemToTerm, sortItems } from '../../../utilities';
 
 const LeadOrganization = ({ handleUpdate }) => {
   const dispatch = useDispatch();
