+++ src/utilities/isEmptyObj.js
@@ -1,0 +1,5 @@
+export const isEmptyObj = objToCheck => {
+  return (
+    Object.entries(objToCheck).length === 0 && objToCheck.constructor === Object
+  );
+};
